django/requirements.txt
django==1.8
django-haystack>=2.4
whoosh
django-debug-toolbar